# Conflicts:
#	framework/src/main/java/org/tron/core/config/args/Args.java
#	framework/src/test/java/org/tron/core/config/args/ArgsTest.java
#	plugins/build.gradle

+++ framework/src/main/java/org/tron/core/zen/address/SpendingKey.java
@@ -89,7 +89,7 @@ public class SpendingKey {
           throw new BadItemException(
               "librustzcash_check_diversifier does not return valid diversifier");
         }
-        blob[33] += 1;
+        blob[33] += (byte) 1;
       } finally {
         JLibsodium.freeState(state);
       }
